minishell power

+++ libft/src/ft_atoi.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int	ft_atoi(const char *nptr)
+long long int	ft_atoi(const char *nptr)
 {
 	long long int	val;
 	int				signe;
